+++ packages/cli/packages/quickHelpers/mergeUx.js
@@ -22,14 +22,6 @@ let needUpdate = (id, originalCode) => {
     return false;
 };
 
-//获取dist路径
-let getDist = (filePath) =>{
-    let dist = utils.updatePath(filePath, config.sourceDir, 'dist');
-    let { name, dir} =  path.parse(dist);
-    let distPath = path.join(dir, `${name}.css`);
-    return distPath;
-};
-
 function beautifyUx(code){
     return beautify.html(code, {
         indent: 4,
@@ -38,17 +30,8 @@ function beautifyUx(code){
 }
 
 function isNodeModulePath(fileId){
-    let isWin = utils.isWin();
-    let reg = isWin ? /\\node_modules\\/ : /\/node_modules\//;
+    let reg = /[\\/]node_modules[\\/]/;
     return reg.test(fileId);
-}
-
-function fixPath(fileId, dep){
-    if (!isNodeModulePath(fileId)) {
-        return path.join(cwd, config.sourceDir, dep);
-    }
-    let retPath = utils.updatePath(dep, 'npm', 'node_modules');
-    return path.join(cwd, retPath); 
 }
 
 function nodeModules2Npm(p) {
@@ -71,13 +54,13 @@ let map = {
                 if (/node_modules/.test(sourcePath)) {
                     sourcePath = nodeModules2Npm(sourcePath);
                 }
-                targetPath = path.join(cwd, 'source/npm/schnee-ui', relativePath);
+                targetPath = path.join(cwd, 'source', relativePath);
             }
             let importSrc = path.relative(
                 path.dirname(sourcePath),
                 targetPath
             );
-            importSrc = utils.isWin() ? importSrc.replace(/\\/g, '/'): importSrc;
+            importSrc = importSrc.replace(/\\/g, '/');
             importTag += `<import name="${i}" src="${importSrc}.ux"></import>`;
         });
         return importTag;
@@ -88,15 +71,15 @@ let map = {
         return `<script>\n${code}\n</script>`;
     },
     resolveComponents: function(data, queue){
-        let { result, sourcePath, relativePath } = data;
-        let isComponentReg = utils.isWin() ? /\\components\\/ : /\/components\// ;
+        let {result, sourcePath, relativePath} = data;
+        let isComponentReg = /[\\/]components[\\/]/;
         if (!isComponentReg.test(sourcePath)) return;
         queue.push({
             code: beautify.js(result.code.replace('console.log(nanachi)', 'export {React}')),
             path: relativePath,
             type: 'js'
         });
-        let reg = utils.isWin() ? /components\\(\w+)/ :  /components\/(\w+)/;
+        let reg = /components[\\/](\w+)/;
         var componentName =  sourcePath.match(reg)[1];
         result.code = `import ${componentName}, { React } from './index.js';
         export default  React.registerComponent(${componentName}, '${componentName}');`;
@@ -104,9 +87,10 @@ let map = {
 
 };
 
-module.exports = (data, queue)=>{
-    let { sourcePath, result, relativePath } = data;
-    var uxFile = quickFiles[utils.fixWinPath(sourcePath)];
+module.exports = async (data, queue)=>{
+    let {sourcePath, result} = data;
+    sourcePath = utils.fixWinPath(sourcePath);
+    var uxFile = quickFiles[sourcePath];
     //如果没有模板, 并且不是app，则认为这是个纯js模块。
     if (!uxFile || (!uxFile.template && uxFile.type != 'App')) {
         return {