+++ packages/cli/packages/utils/mergeConfigJson.js
@@ -1,6 +1,9 @@
 const path = require('path');
 const buildType = process.env.ANU_ENV;
 
+
+
+
 module.exports = function(modules, json) {
     
     if (modules.componentType !== 'App') {
@@ -8,14 +11,15 @@ module.exports = function(modules, json) {
     }
     let configJson = {};
     try {
-        configJson = require( path.join(process.cwd(), `${buildType}Config.json` ));
+        configJson = require( path.join(process.cwd(), 'source', `${buildType}Config.json` ));
     } catch (err) {
        
     }
 
-    delete configJson.subPackages;
-    delete configJson.subpackages;
-
+    if (buildType != 'quick') {
+        delete configJson.subPackages;
+        delete configJson.subpackages;
+    }
     Object.assign(json, configJson);
     return json;
 }