+++ packages/cli/packages/utils/setSubPackage.js
@@ -2,8 +2,9 @@
 //分包配置
 const path = require('path');
 const buildType = process.env.ANU_ENV;
-const supportPlat = ['wx', 'bu', 'qq'];
+const supportPlat = ['wx', 'bu', 'qq', 'ali'];
 const keys = {
+    ali: 'subPackages',
     bu: 'subPackages',
     wx: 'subpackages',
     qq: 'subpackages'
@@ -22,7 +23,7 @@ module.exports = function(modules, json) {
     json[keys[buildType]] = json[keys[buildType]] || [];
     let subPackages = [];
     try {
-        appRootConfig = require( path.join(process.cwd(), `${buildType}Config.json`) );
+        appRootConfig = require( path.join(process.cwd(), 'source', `${buildType}Config.json`) );
         subPackages = appRootConfig.subpackages || appRootConfig.subPackages || [];
     } catch (err) {
        
@@ -51,6 +52,11 @@ module.exports = function(modules, json) {
             name: name,
             pages:[]
         };
+
+        if (buildType === 'ali') {
+            delete subPackagesItem.name;
+        }
+
         //核心是根据配置中的 resource 创建正则，去遍历出 pages 中的的匹配这个正则的路由。
         let reg = new RegExp('^' + resource);
         json[keys[buildType]].push(subPackagesItem);