+++ packages/render/miniapp/render.all.js
@@ -1,7 +1,7 @@
 import { isFn, toLowerCase, get } from 'react-core/util';
 import { createRenderer } from 'react-core/createRenderer';
 import { render } from 'react-fiber/scheduleWork';
-import { updateMiniApp, _getApp, delayMounts, registeredComponents } from './utils';
+import { updateMiniApp, _getApp, delayMounts } from './utils';
 
 var onEvent = /(?:on|catch)[A-Z]/;
 
@@ -17,9 +17,9 @@ export let Renderer = createRenderer({
         let { props, lastProps } = fiber;
         let beaconId = props['data-beacon-uid'];
         let instance = fiber._owner; //clazz[instanceId];
-        if (instance && !instance.classUid) {
-            instance = get(instance)._owner;
-        }
+       // if (instance && !instance.classUid) {
+       //     instance = get(instance)._owner;
+       // }
         if (instance && beaconId) {
             var cached =
                 instance.$$eventCached || (instance.$$eventCached = {});