+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,59 @@
 # Changelog
 
+# 1.2.6 (2019-05-27)
+## CLI
+#### Bug fix
+- 修复快应用补丁组件引用问题
+- 修复静态资源别名解析路径bug
+
+# 1.2.5 (2019-05-24)
+
+## 核心库
+#### Feature
+- 抹平小程序与快应用在页面组件的onTabItemTap差异。
+- 组件支持静态的options对象， 以支持微信小程序的自定义组件的options.styleIsolation功能。 https://developers.weixin.qq.com/miniprogram/dev/framework/custom-component/wxml-wxss.html
+
+#### Bug fix
+- 修正onShare在页面只触发一次的BUG https://github.com/RubyLouvre/anu/issues/1000
+
+## CLI
+#### Feature
+- 请允许components出现在pages下面，以实现更好的分包功能。https://rubylouvre.github.io/nanachi/documents/subpackages.html
+- 重构qunar样板工程，以演示分包功能。
+
+
+# 1.2.4 (2019-05-17)
+
+## 核心库
+#### Feature
+- 添加React Hooks的支持，添加箭头函数的支持。
+- ReactWX添加结useState, useContext, useEffect的支持，并去掉Children, createPortal, cloneElement, isValidElement。
+- 抹平微信和qq小程序中getStorage api差异。
+
+#### Bug fix
+- 事件里可以条件绑定函数。
+
+## CLI
+#### Feature
+- regenerator-runtime锁定版本号。
+- 支持快应用自定义合并router。
+
+#### Bug fix
+- windows快应用产物后缀名bug。 [#984](https://github.com/RubyLouvre/anu/issues/984)
+
+# 1.2.2 (2019-04-26)
+
+## CLI
+#### Feature
+- 对 app.js 中 是否有 globalData 对象做强制校验。
+- 支持 QQ 小程序分包。
+
+#### Bug fix
+- 兼容快应用中 css 动画规则 keyframes。
+- 修复快应用下 showToast bug。
+- 修复静态资源 copy 路径 bug。
+
+
 
 # 1.2.1 (2019-04-19)